SMB2: set session server port in partial capture
When dissecting a capture that does not have a session setup exchange,
assume tree connect response packets are coming from the server and set
the port accordingly.

This is needed for decryption because the decryption key is different
depending on which side the message was created on. We need to set the
server port to know if a packet was made on the server.

Change-Id: I80229ed6d69821bc0c76808e57bdb2fc53fcceb3
Reviewed-on: https://code.wireshark.org/review/21741
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>