WSDG: trivial: fix a link
Asciidoc will just create a link from a URL that appears in the text. If
we want the link to appear as an address without a description, there's
no need to add an empty description. We should, however, have a space
behind the link to separate it from the text that follows.

Change-Id: Ic01b1dbbea922d785776f1965481ed58f389e30a
Reviewed-on: https://code.wireshark.org/review/21789
Reviewed-by: Martin Kaiser <wireshark@kaiser.cx>
Petri-Dish: Martin Kaiser <wireshark@kaiser.cx>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>