Merge with FreeRADIUS and update.
Pick up the copyright notice and $Id value.

Don't double-comment-out the items we handle in
packet-radius_packetcable.c; they're already commented out in the
FreeRADIUS version, and that reduces the number of differences between
us and them.  Note in a comment before them that we're handling them in
packet-radius_packetcable.c.

Add a comment pointing at various CableLabs specs before the attributes,
noting where in the specs they're mentioned. (Hint: NOT in a section
about RADIUS!  I guess the idea is that some protocol *other* than
RADIUS can be used to transfer them.)

Change-Id: Ib16fb10df9a434f79a05ce0537c671076aff11a2
Reviewed-on: https://code.wireshark.org/review/21810
Reviewed-by: Guy Harris <guy@alum.mit.edu>