ECDSA+DSA SHA2 algorithm OIDs added to PKCS1 dissector
Those algorithms are e.g. used to sign X.509 certificates, or in CMP (RFC4210)
messages.

It is not fully deterministic to which dissector those OIDs should be added,
PKCS1 is chosen as there are already the OIDs for the ECC curves and the OIDs
in question are also listed in PKIXAlgs-2009.asn.

Change-Id: I4fc41d27220400c37695ecce5d7f05726ad39af2
Reviewed-on: https://code.wireshark.org/review/22049
Petri-Dish: Pascal Quantin <pascal.quantin@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Pascal Quantin <pascal.quantin@gmail.com>