Handle filter open and read errors better.
If the user's color filter file exists but can't be opened, fail and
return a failure indication.

If it doesn't exist, just try reading the global color filter file.

If *that* exists but can't be opened, fail and return a failure
indication; if it doesn't exist, don't treat that as an error.

Change-Id: I5ebdc0eb1ada3cc190990be9228639304f32147e
Reviewed-on: https://code.wireshark.org/review/22054
Reviewed-by: Guy Harris <guy@alum.mit.edu>