Update to newer versions of libgcrypt and GnuTLS.
We want a newer version of libgcrypt to get additional crypto functions.
Update to the current release, 1.7.7.

Update to the current release of libgpg-error, 1.27, while we're at it.

Update to the current "stable" version of GnuTLS, 3.4.17; 2.12.19
doesn't work with libgcrypt 1.7.7.  3.4.17 requires Nettle, and Nettle
requires GMP, so, if we're building with GnuTLS, download and install
the current versions of Nettle and GMP.

GMP requires lzip, so download and install it as well.

Clean up some "version >= x.y.z" checks to check the major version
number in all cases.

Change-Id: I39cccd34e0d7f49ac35b0bbacdab03251d42a1de
Reviewed-on: https://code.wireshark.org/review/22156
Reviewed-by: Guy Harris <guy@alum.mit.edu>