Put the {un}install_XXX() definitions in the right order.
Put them in the same order as the order in which the _VERSION values are
defined and in which they're invoked.

Similarly, do the "make sure we have the requested version installed"
tests in the same order as the un-installation order (which is the
reverse of the installation order).

Change-Id: I0e2bd1d249832090c3d81bacfe010de19de54cdf
Reviewed-on: https://code.wireshark.org/review/22158
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit d01e9df47258cdd6bf66ace1d6f592773b574926)
Reviewed-on: https://code.wireshark.org/review/22999