USB CCID: register obosolete preference
In 082e3e346f69f0c2134064e80dcc104c18111c55, we dropped the prtype
preference in favour of Descode As.

Register prtype as an obsolete preference to make sure that it's not
removed from the preferences file. The way, the preferences file is
still usable with older wireshark versions.

Change-Id: I8feed6080b58dd5443898e2c5b12732b0b3a0a4f
Reviewed-on: https://code.wireshark.org/review/22373
Petri-Dish: Martin Kaiser <wireshark@kaiser.cx>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Martin Kaiser <wireshark@kaiser.cx>