Clean up the "non-constant initalizers" stuff.
We aren't, in this branch, insisting on C99 on UN*X, so we still warn
against initializing aggregates, even if they have automatic storage
duration, with non-constant values.  Put that into its own section, with
some details, and take that out of the section on initializing global
and static variables with non-constant values.

(In practice, it'll probably Just Work, but we haven't pulled any of the
*other* C99isms into this branch, and this branch is soon going to be
less active once 2.4 is out, so we'll just leave the restrictions in.)

For the latter section, expand the example in the hopes of avoiding
confusion between "static storage duration" (which something declared
"static" has, but which anything declared with file scope, whether
declared "static" or not, also has) and "static storage duration and
internal linkage", which is what the "static" keyword specifies.

Change-Id: I5a7d307f0cb3be84e355b92231fb115359d0cd41
Reviewed-on: https://code.wireshark.org/review/22436
Reviewed-by: Guy Harris <guy@alum.mit.edu>