Fix crash in UDT dissector when long NAK packet is dissected
Change is to limit the number of entries from the NAK message included in the
summary line (and add ellipsis if there are more than will fit).

In addition, add checks to make sure we dont read beyond the end of the
captured packet when parsing NAKs.

Change-Id: I60db4b62d86c05329eb7c79ae1927eeb1b7e11ba
Reviewed-on: https://code.wireshark.org/review/22733
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>