ZigBee fix Default Response dissection
The ZCL Default Response command can be sent in response to any profile-wide or
cluster specific command. The Default Response command is itself a profile-wide
command so the Default Response frame control is of no help in deciding whether
the original request is profile-wide or cluster-specific. The simplest solution
is to not attempt interpretation. A more sophisticated solution would be to cache
all ZCL frame counters and match Default Responses based on sequence number, but
this is problematic because sequence numbers repeat. At least for now we can
always display the information correctly.

Change-Id: I827e2d2f9d6e5f7c9dfa572d2ee2ac3c9f170d70
Reviewed-on: https://code.wireshark.org/review/22688
Reviewed-by: Anders Broman <a.broman58@gmail.com>