Clean up configuration file and plugin folder description.
Have a separate section, before all the others, giving the top-level
directories under which those folders are placed, with names for use in
later sections.

Update the example personal application data folder on Windows to the NT
6.x-and-beyond standard.

Don't give the pathnames for configuration files in the table, just give
the name and the description.

For the global files, distinguish between macOS and other Unix-like
systems, and, for the latter, mention both /usr/XXX and /usr/local/XXX.

The preferences file isn't "wireshark.conf", and hasn't been that for
quite a while.

For all the configuration files, give the details of personal vs. global
files - they're different for different files.

Have separate sections for configuration files and plugin folders.  For
plugin folders, note both the use of the top-level plugin folder for Lua
scripts and the use of the per-Wireshark-version subfolder for compiled
plugins.

Use fixed-format text for the non-variable parts of pathnames, and
italics for the variable parts.

This should, among other things, make it easier for other documents,
such as the Lua documentation in the Wireshark wiki, to refer to folders
such as the configuration file and plugin folders.

Change-Id: I133c1e159e992827458bee64c4f37be5b50f9b6f
Reviewed-on: https://code.wireshark.org/review/23060
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit de3a563d6698de8e45708e9d056f948c8eab7c8c)
Reviewed-on: https://code.wireshark.org/review/23067