Update RFC and I-D references.
Give draft-frascone-xml-dictionary for the dictionary syntax.

Give RFC numbers and I-D names, rather than URLs to the plain text
versions of them; that way, you can more easily go to the HTML versions,
which give more information.

The "more information" for I-Ds may include later drafts and the final
RFCs, and for RFCs may give newer RFCs obsoleting the older ones; add
the RFC numbers for the I-Ds and the RFC number for the current Diameter
RFC.

Change-Id: If01ea341af3ae892755a243bd1dd66acfdfd5062
Reviewed-on: https://code.wireshark.org/review/23086
Reviewed-by: Guy Harris <guy@alum.mit.edu>